APP/requirements.txt
uvicorn == 0.20.0
fastapi==0.68.0
numpy== 1.24.2
psycopg2-binary
Jinja2==3.0.3
SQLAlchemy==2.0.25
sqlalchemy-orm==1.2.10
requests==2.28.2

